fix: request #41851 Initial effort value should not be rounded to integer
How to test:
- Set up a User Stories tracker with a Float field for the Initial
effort semantic.
- Create two stories with floating values for initial effort, for
example "0.5" and "1.5"
- Go to the Top Backlog in the same project.
- Plan those stories in a Release.
- The round badge next to the story cards should be the floating value,
it should not be rounded to the integer.
- The square badge in the Release that sums up all the initial efforts
should have the correct value ("2" instead of "1").

No functional change expected for initial effort with list of static
numeric values, computed and integer fields.

Introduced by dd7b60ab4e3a5546050e99a7a7af8a794aad44d0

+++ plugins/agiledashboard/include/AgileDashboard/Milestone/Backlog/IBacklogItem.php
@@ -32,9 +32,9 @@ interface IBacklogItem
 {
     public const REMAINING_EFFORT_FIELD_NAME = 'remaining_effort';
 
-    public function setInitialEffort(?int $value): void;
+    public function setInitialEffort(?float $value): void;
 
-    public function getInitialEffort(): ?int;
+    public function getInitialEffort(): ?float;
 
     public function setStatus(string $status, string $status_semantic): void;
